pyproject.toml
[tool.poetry]
name = "fastapi-courses"
version = "0.1.0"
description = ""
authors = ["robond999 <robert.aydinyan@esterox.am>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
sqlalchemy = "^2.0.40"
uvicorn = "^0.34.2"
asyncpg = "^0.30.0"
python-jose = "^3.4.0"
python-dotenv = "^1.1.0"
pydantic = {extras = ["email"], version = "^2.11.4"}
requests = "^2.32.3"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
httpx = "^0.28.1"
python-multipart = "^0.0.20"
alembic = "^1.16.1"
psycopg2-binary = "^2.9.10"
pytest = "^8.4.0"
pytest-asyncio = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
